Moved liquid cooled battery and motor to battery and motor files and broke apart remaining components from heat sink file because it did not have a common theme
