+++ openconcept/thermal/__init__.py
@@ -2,14 +2,10 @@ from .chiller import HeatPumpWithIntegratedCoolantLoop
 from .ducts import ExplicitIncompressibleDuct, ImplicitCompressibleDuct, ImplicitCompressibleDuct_ExternalHX
 from .heat_exchanger import HXGroup
 from .heat_pipe import HeatPipe
-from .heat_sinks import (
-    BandolierCoolingSystem,
-    LiquidCooledBattery,
-    MotorCoolingJacket,
-    LiquidCooledMotor,
-    SimplePump,
-    SimpleHose,
-)
+from .battery_cooling import BandolierCoolingSystem
+from .motor_cooling import MotorCoolingJacket
+from .pump import SimplePump
+from .hose import SimpleHose
 from .manifold import FlowSplit, FlowCombine
 from .thermal import (
     PerfectHeatTransferComp,
